add clock frequency determination
It turns out that you can, sort of, get CPU frequency information
from leaf 0x15, if that's present. If it's not present, you can
parse the information out of the brand string -- Intel actually
documented this as the Official Correct Way to get CPU speed
for quite a while, complete with elaborate parsing logic.

Example:

https://www.scss.tcd.ie/~jones/CS4021/processor-identification-cpuid-instruction-note.pdf

This patch tries to use leaf 0x15 if it's present, which may not
be the wisest choice -- there's rumors of it being buggy on some
parts, but I don't know details. If it's not present, or it would
produce 0 or a division by 0 error, we try the brand string.
If we fail, we yield -1. Running this against the mock CPU tests
revealed what look like reasonably plausible values.