workflows: Bump golang versions
Let's make sure we're doing the release using the latest golang version,
and that we're testing against the supported stable versions.

Please, refer to https://endoflife.date/go for more info about the
golang versions that are still supported.

NOTE: We're keeping 1.19.x as @klauspost tests with the latest EOL
release as well.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>